feat(zsh): remove zoxide
This commit is made as part of the effort to simplify and minimalize my
environment. Zoxide can be replaced with `cd`, and frequently visited
directories can be saved into sessions with `tmux`. Custom scripts can
be written in the event that quick directory switching is still
required, even after all this.