feat(zsh): remove eza
This commit is made as part of the effrot to simplify and minimalize my
environment. Eza can be replaced with `ls`, as there is no need for any
special file listing formatting, and it is generally better to learn
how to use the most standard tools for adaptability to any environment.